Move the resampling code out into its own module so that it can be reused
