+++ emission/analysis/intake/cleaning/clean_and_resample.py
@@ -20,7 +20,6 @@ from builtins import *
 from past.utils import old_div
 import logging
 import numpy as np
-import scipy.interpolate as spi
 import pandas as pd
 import arrow
 import geojson as gj
@@ -28,6 +27,8 @@ import json
 
 # Our imports
 import emission.analysis.config as eac
+import emission.analysis.intake.location_utils as eail
+
 import emission.storage.decorations.timeline as esdtl
 import emission.storage.decorations.trip_queries as esdtq
 import emission.storage.decorations.analysis_timeseries_queries as esda
@@ -411,7 +412,25 @@ def get_filtered_points(section, filtered_section_data):
     logging.debug("Resampling data of size %s" % len(filtered_loc_df))
     # filtered_loc_list has removed the outliers. Now, we resample the data at
     # 30 sec intervals
-    resampled_loc_df = resample(filtered_loc_df, interval=30)
+
+    """
+    TODO: There is a problem with working on this on a section by section basis.
+    Namely, how do we deal with the ends. In particular, note that our sections
+    are generated by looking at motion activity while our points are generated
+    by looking at locations, and the two streams are not synchronized. So we have
+    "transition areas" between sections that have no points. Resampling will
+    help with this because we can ensure that arbitrary time point - for example,
+    the transition between sections, has an associated location point. But that
+    fix really needs to be made in the segmentation code - by this point, the
+    original timestamps have been replaced by the timestamps of the start and
+    end location points. But we can also have a gap if the duration of the section
+    is not a multiple of the interval, as it is likely to be. In this case, we will
+    manually add the last entry to ensure that we adequately capture the end point.
+    :param filtered_loc_df:
+    :param interval:
+    :return:
+    """
+    resampled_loc_df = eail.resample(filtered_loc_df, interval=30)
 
     with_speeds_df = eaicl.add_dist_heading_speed(resampled_loc_df)
     with_speeds_df["idx"] = np.arange(0, len(with_speeds_df))
@@ -729,110 +748,6 @@ def remove_outliers(raw_loc_entry_list, filtered_point_id_list):
                   (len(loc_df), len(valid_loc_df),
                    len(loc_df.columns), len(filtered_loc_df.columns)))
     return filtered_loc_df
-
-def resample(filtered_loc_df, interval):
-    """
-    TODO: There is a problem with working on this on a section by section basis.
-    Namely, how do we deal with the ends. In particular, note that our sections
-    are generated by looking at motion activity while our points are generated
-    by looking at locations, and the two streams are not synchronized. So we have
-    "transition areas" between sections that have no points. Resampling will
-    help with this because we can ensure that arbitrary time point - for example,
-    the transition between sections, has an associated location point. But that
-    fix really needs to be made in the segmentation code - by this point, the
-    original timestamps have been replaced by the timestamps of the start and
-    end location points. But we can also have a gap if the duration of the section
-    is not a multiple of the interval, as it is likely to be. In this case, we will
-    manually add the last entry to ensure that we adequately capture the end point.
-    :param filtered_loc_df:
-    :param interval:
-    :return:
-    """
-    loc_df = filtered_loc_df
-    # See https://github.com/e-mission/e-mission-server/issues/268 for log traces
-    #
-    # basically, on iOS, due to insufficient smoothing, it is possible for us to
-    # have very small segments. Some of these contain zero points, and we skip them
-    # in the segmentation stage. Some of them contain one point, and we don't.
-    # Ideally, we would strip these sections too and merge the stops on the two sides
-    # But that is going to take more time and effort than I have here.
-    #
-    # So let's just return the one point without resampling in that case, and move on for now
-    if len(loc_df) == 0 or len(loc_df) == 1:
-        logging.debug("len(loc_df) = %s, early return" % (len(loc_df)))
-        return loc_df
-
-    logging.debug("Resampling entry list %s of size %s" %
-                  (loc_df[["fmt_time", "ts", "longitude", "latitude"]].head(), len(filtered_loc_df)))
-    start_ts = loc_df.ts.iloc[0]
-    end_ts = loc_df.ts.iloc[-1]
-    tz_ranges_df = _get_tz_ranges(loc_df)
-    logging.debug("tz_ranges_df = %s" % tz_ranges_df)
-
-    lat_fn = spi.interp1d(x=loc_df.ts, y=loc_df.latitude, bounds_error=False,
-                          fill_value='extrapolate')
-
-    lng_fn = spi.interp1d(x=loc_df.ts, y=loc_df.longitude, bounds_error=False,
-                          fill_value='extrapolate')
-
-    altitude_fn = spi.interp1d(x=loc_df.ts, y=loc_df.altitude, bounds_error=False,
-                          fill_value='extrapolate')
-
-    ts_new = np.append(np.arange(start_ts, end_ts, 30), [end_ts])
-    logging.debug("After resampling, using %d points from %s -> %s" %
-                  (ts_new.size, ts_new[0], ts_new[-1]))
-    lat_new = lat_fn(ts_new)
-    lng_new = lng_fn(ts_new)
-    alt_new = altitude_fn(ts_new)
-    tz_new = [_get_timezone(ts, tz_ranges_df) for ts in ts_new]
-    ld_new = [esdl.get_local_date(ts, tz) for (ts, tz) in zip(ts_new, tz_new)]
-    loc_new = [gj.Point((lng, lat)) for (lng, lat) in zip(lng_new, lat_new)]
-    fmt_time_new = [arrow.get(ts).to(tz).isoformat() for
-                        (ts, tz) in zip(ts_new, tz_new)]
-    loc_df_new = pd.DataFrame({"latitude": lat_new, "longitude": lng_new,
-                               "loc": loc_new, "ts": ts_new, "local_dt": ld_new,
-                               "fmt_time": fmt_time_new, "altitude": alt_new})
-    return loc_df_new
-
-def _get_timezone(ts, tz_ranges_df):
-    if len(tz_ranges_df) == 1:
-        sel_entry = tz_ranges_df
-    else:
-        # TODO: change this to a dataframe query instead
-        sel_entry = tz_ranges_df[(tz_ranges_df.start_ts <= ts) &
-                            (tz_ranges_df.end_ts >= ts)]
-    if len(sel_entry) != 1:
-        logging.warning("len(sel_entry) = %d, using the one with the bigger duration" % len(sel_entry))
-        sel_entry["duration"] = sel_entry.end_ts - sel_entry.start_ts
-        sel_entry = sel_entry[sel_entry.duration == sel_entry.duration.max()]
-    return sel_entry.timezone.iloc[0]
-
-def _get_tz_ranges(loc_df):
-    tz_ranges = []
-    if len(loc_df) == 0:
-        logging.debug("Called with loc_df of length 0, returning empty" % len(loc_df))
-        return tz_ranges
-
-    # We know that there is at least one entry, so we can access it with impunity
-    curr_start_ts = loc_df.ts.iloc[0]
-    curr_tz = loc_df.local_dt_timezone.iloc[0]
-    for row in loc_df.to_dict('records'):
-        loc_data = ad.AttrDict(row)
-        if loc_data.local_dt_timezone != curr_tz:
-            tz_ranges.append({'timezone': curr_tz,
-                              'start_ts': curr_start_ts,
-                              'end_ts': loc_data.ts})
-            curr_start_ts = loc_data.ts
-            curr_tz = loc_data.local_dt_timezone
-
-    # At the end, always add an entry
-    # For cases in which there is only one timezone (common case),
-    # this will be the only entry
-    tz_ranges.append({'timezone': curr_tz,
-                      'start_ts': curr_start_ts,
-                      'end_ts': loc_df.ts.iloc[-1]})
-    logging.debug("tz_ranges = %s" % tz_ranges)
-    return pd.DataFrame(tz_ranges)
 
 def link_trip_timeline(tl, section_map, stop_map):
     filled_in_sections = [_fill_section(s, section_map[s.get_id()], stop_map) for s in tl.trips]
